requirements.txt
streamlit
google-generativeai
selenium
webdriver-manager